backend/pyproject.toml

[tool.poetry]

name = "dating-app-backend"
version = "0.1.0"
description = "Backend for a dating app built with FastAPI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = {extras = ["standard"], version = "^0.121.0"}
uvicorn = "^0.38.0"
sqlalchemy = "^2.0.44"
pydantic = "^2.12.4"
python-multipart = "^0.0.20"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.2.12"}
python-dotenv = "^1.2.1"
alembic = "^1.17.1"
pydantic-settings = "^2.12.0"
python-jose = "^3.3.0"
greenlet = "^3.2.4"
bcrypt = "<4.0"
boto3 = "^1.35.0"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.2"
pytest-dotenv = "^0.5.2"
requests = "^2.32.5"

[tool.pytest.ini_options]
pythonpath = ["app"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

